refactor(setup): split setup command by responsibility
